Convert CalcHistogram to a compute kernel, because some Macs apparently don't like a target-less fragment shader (?)
FFCC causes Photon Transfer to crash in -[id<MTLDevice> newRenderPipelineStateWithDescriptor:error:] on a x86_64 iMac16,2 running macOS 12.7.6, complaining that "No valid pixelFormats set.".

The MTLRenderPipelineDescriptor.colorAttachments[0].pixelFormat is intentionally `MTLPixelFormatInvalid` though because CalcHistogram's fragment shader is target-less (ie it returns void, which is valid according to the Metal Shader Language PDF.) It's not clear why this scheme works on some Macs but fails on this particular one, but we'll just work around it by converting CalcHistogram into a compute kernel.